+++ lib/index.d.ts
@@ -6,3 +6,4 @@ export * from "./request";
 export * from "./response";
 export * from "./router";
 export * from "./utils";
+export * from "./logger";